requirements.txt
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.0.1
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
